// projects/sivimss-gui/src/app/modules/salas/models/salas.interface.ts
export interface Sala {
  id?: number;
  numEmpleado?: string;
  curp?: string;
  nombre?: string;
  primerApellido?: string;
  segundoApellido?: string;
  fechaNacimiento?: string;
  fechaIngreso?: string;
  sueldoBase?: string;
  velatorio?: string;
  categoria?: string;
  antiguedad?: string;
  correo?: string;
  puesto?: string;
  diasDescanso?: string;
  nombreSala?: string;
  descripcion?: string;
  estatus?: boolean;
  // costoInicial?: string;
  // costoReferencia?: string;
  // precio?: string;
  // region?: string;
  // clave?: string;
  // servicios?: Servicio[];
  // articulos?: Articulo[];
}